testTzSys: fix test suite for non-glibc when using TZDIR
If TZDIR was set to a directory differing from the system timezone
directory, the testTZSys test suite would fail on certain timezones
in certain configurations. We discovered this when building the tz
packaged in nixpkgs on macOS: We would provide our packaged tzdata via
TZDIR, but /usr/share/zoneinfo would still be accessible. Since Apple's
tzdata seem to be outdated, the tests fail consistently. This is because
the tzset(3) implementation of Apple's libSystem is “only”
POSIX-conforming and doesn't support the glibc invention of TZDIR. They
are not alone with this -- musl doesn't support it as well.

To work around this we can use POSIX behavior, however: If TZ is
prefixed with a ':', we can pass an absolute path and rely on the lookup
behavior implemented in tz itself.